thor: Move CpuData-related structs to their own header
