+++ kernel/thor/arch/riscv/thor-internal/arch/cpu.hpp
@@ -7,6 +7,7 @@
 #include <frg/tuple.hpp>
 #include <frg/vector.hpp>
 #include <initgraph.hpp>
+#include <thor-internal/arch-generic/cpu-data.hpp>
 #include <thor-internal/arch/ints.hpp>
 #include <thor-internal/arch/unimplemented.hpp>
 #include <thor-internal/error.hpp>
@@ -218,78 +219,6 @@ private:
 };
 
 size_t getStateSize();
-
-// Note: These constants we mirrored in assembly.
-// Do not change their values!
-inline constexpr unsigned int uarRead = 1;
-inline constexpr unsigned int uarWrite = 2;
-
-// Note: This struct is accessed from assembly.
-// Do not change the field offsets!
-struct UserAccessRegion {
-	void *startIp;
-	void *endIp;
-	void *faultIp;
-	unsigned int flags;
-};
-
-struct IseqContext;
-
-// Note: This struct is accessed from assembly.
-// Do not change the field offsets!
-struct AssemblyCpuData {
-	AssemblyCpuData *selfPointer; //  0x0
-	uint64_t currentDomain;       //  0x8
-	void *exceptionStackPtr;      // 0x10
-	void *irqStackPtr;            // 0x18
-	uint64_t scratchSp;           // 0x20
-	UserAccessRegion *currentUar; // 0x28
-	IseqContext *iseqPtr;
-};
-
-inline void writeToTp(AssemblyCpuData *context) { asm volatile("mv tp, %0" : : "r"(context)); }
-
-struct Thread;
-
-struct PlatformCpuData : public AssemblyCpuData {
-	// Bits of the pendingIpis field.
-	// Since RISC-V only has a single IPI vector, we need to emulate multiple IPIs in software.
-	static constexpr uint64_t ipiPing = UINT64_C(1) << 0;
-	static constexpr uint64_t ipiShootdown = UINT64_C(1) << 1;
-	static constexpr uint64_t ipiSelfCall = UINT64_C(1) << 2;
-
-	uint64_t hartId{~UINT64_C(0)};
-
-	// Executor image that we use to save/restore state.
-	Executor *activeExecutor{nullptr};
-
-	// Actual value of the FS field in sstatus before it was cleared in the kernel.
-	// Zero (= extOff) indicates that the current register state cannot be relied upon
-	// (i.e., it has been saved to the executor or it is in control of U-mode).
-	uint8_t stashedFs{0};
-
-	std::atomic<uint64_t> pendingIpis;
-
-	// Deadlines for global timers and preemption.
-	// frg::null_opt is the respective deadline is not set (= infinite).
-	frg::optional<uint64_t> timerDeadline;
-	frg::optional<uint64_t> preemptionDeadline;
-	// Current deadline programmed into the supervisor timer.
-	uint64_t currentDeadline{~UINT64_C(0)};
-
-	UniqueKernelStack irqStack;
-
-	frg::manual_box<AsidCpuData> asidData;
-
-	uint32_t profileFlags = 0;
-};
-
-// Get a pointer to this CPU's PlatformCpuData instance.
-inline PlatformCpuData *getPlatformCpuData() {
-	AssemblyCpuData *result;
-	asm volatile("mv %0, tp" : "=r"(result));
-	return static_cast<PlatformCpuData *>(result);
-}
 
 // Determine whether this address belongs to the higher half.
 inline constexpr bool inHigherHalf(uintptr_t address) {
